requirements.txt
fastapi
uvicorn
python-dotenv
requests
yfinance
faiss-cpu
sentence-transformers
transformers
peft
torch
accelerate
bitsandbytes
newsapi-python
